CHORE - Logic components that are not in the logic folder have been deleted and renamed folders to lower case
