--- src/algorithms/epsilonGreedy.test.js
@@ -1,10 +1,0 @@
-import * as EpsilonGreedy from './epsilonGreedy.js';
-import { describe } from 'vitest';
-import { test, expect } from 'vitest';
-
-describe('epsilonGreedy.js', () => {
-
-    test('Vitest läuft', () => {
-        expect(true).toBe(true);
-    });
-});
