--- src/algorithms/greedy.test.js
@@ -1,10 +1,0 @@
-import * as Greedy from './greedy';
-import { describe } from 'vitest';
-import { test, expect } from 'vitest';
-
-describe('greedy.js', () => {
-
-    test('Vitest läuft', () => {
-        expect(true).toBe(true);
-    });
-});
