+++ src/logic/CurrentGame.test.js
@@ -1,6 +1,6 @@
 import { describe, it, expect, beforeEach } from 'vitest';
-import { DistributionTyp } from './Enumeration/DistributionTyp.js';
-import { AlgorithmTyp } from './Enumeration/AlgorithmTyp.js';
+import { DistributionTyp } from '@/logic/enumeration/DistributionTyp.js';
+import { AlgorithmTyp } from '@/logic/enumeration/AlgorithmTyp.js';
 import CurrentGame from './CurrentGame.js';
 
 describe('CurrentGame', () => {
