+++ src/logic/algorithm/Algorithm.js
