+++ src/logic/enumeration/AlgorithmTyp.js
