+++ src/logic/enumeration/DistributionTyp.js
